Update README
There is no tesseract-ocr-deu-frak for Tesseract 4.0.
It was replaced by tesseract-ocr-frk.

There is still a tesseract-ocr-deu-frak for Tesseract 3.04,
but it uses a different installation path and does not support LSTM.

In addition to the language models, there are now also script models.

Fixing headers is no longer needed with latest Debian packages and should
also be fixed for Ubuntu soon.

Signed-off-by: Stefan Weil <sw@weilnetz.de>