[WIP] Implement cropping with tesseract
Experiments indicate a double preprocessing effect: Taking the
rectangle defined by the maximum extent of all regions identified
in the 1st pass, using it as the print space and restricting all
subsequent operations (i.e. 2nd pass) to this print space,
increases the quality of both layout and text recognition.