segment*/recognize: workaround for buggy Tesseract page iterators…
Tesseract's `PageIterator` / `ResultIterator` navigation has 2 bugs
which makes them hardly usable: they are not equipped to cleanly
handle the case when
- a non-text block is entered at the PARA/TEXTLINE/WORD level, or
- an empty word (only rejections) is entered at the SYMBOL level.
In particular, `IsAtFinalElement` and `Empty` prematurely signal
True at the lower (but not higher!) levels, the latter w.r.t. the
current and the former w.r.t. the next segment.
Using the API in a naïve / straightforward way would cause all
follow-up segments to be lost, respectively.
This contains a workaround on the API level.